gasops_backend_ai_fabric/requirements.txt
# Database connectivity
pyodbc==5.3.0
azure-identity==1.25.1

# Azure OpenAI
openai>=1.0.0
python-dotenv>=1.0.0

# FastAPI and web framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Utilities
tzdata>=2023.3

# AI Search
azure-search-documents==11.6.0

requests==2.32.5
requests-pkcs12==1.24
azure-ai-documentintelligence==1.0.0